create Plot component and reorganize folders
