+++ love/src/components/GeneralPurpose/Plot/VegaTimeSeriesPlot/TimeSeriesConfig/TSCEntry/TSCEntry.module.css
