+++ love/src/components/GeneralPurpose/Plot/VegaTimeSeriesPlot/TimeSeriesConfig/TSCEntry/TSCInput/TSCInput.module.css
