+++ love/src/components/GeneralPurpose/Plot/VegaTimeSeriesPlot/VegaLegend.jsx
